+++ v3/crates/graphql/ir/src/query_root/select_many.rs
@@ -89,6 +89,7 @@ pub fn select_many_generate_ir<'n, 's>(
                                 argument,
                                 &model_source.type_mappings,
                                 &model_source.data_connector,
+                                session_variables,
                                 &mut usage_counts,
                             )?;
 
@@ -116,6 +117,7 @@ pub fn select_many_generate_ir<'n, 's>(
                     argument.value.as_object()?,
                     &model_source.data_connector,
                     &model_source.type_mappings,
+                    session_variables,
                     &mut usage_counts,
                 )?);
             }
@@ -158,7 +160,7 @@ pub fn select_many_generate_ir<'n, 's>(
         model_source,
         model_arguments,
         query_filter,
-        permissions::get_select_filter_predicate(field_call)?,
+        permissions::get_select_filter_predicate(&field_call.info)?,
         limit,
         offset,
         order_by,